- remove the App folder / namespace
