+++ config/charts.php
